chore: Rename 'searchAndSort' folder as 'search'
